renaming generative replace plugin
