+++ packages/url-loader/src/plugins/replace.ts
