+++ packages/url-loader/tests/plugins/replace.spec.js
@@ -1,8 +1,8 @@
 import { Cloudinary } from "@cloudinary/url-gen";
 
-import * as generativeReplacePlugin from "../../src/plugins/generative-replace";
+import * as replacePlugin from "../../src/plugins/replace";
 
-const { plugin } = generativeReplacePlugin;
+const { plugin } = replacePlugin;
 
 const cld = new Cloudinary({
   cloud: {
